Project list now shown as grid rather than row. Added scrollbar and adjusted max size of skill cards
