+++ styles/Projects.module.css
@@ -1,74 +1,10 @@
-/* @media (max-width: 900px) or (max-height: 850px) {
-   #projectList {
-      display: flex;
-      flex-direction: column;
-      gap: 40px;
-      width: fit-content;
-      height: 350px;
-      margin-left: auto;
-      margin-right: auto;
-      overflow-x: hidden;
-   }
-
-   .projectCard {
-      display: flex;
-      flex-direction: column;
-      flex-wrap: wrap;
-      flex-shrink: 0;
-      font-size: clamp(11px, 1dvw, 12px);
-      width: 765;
-      height: fit-content;
-      border-radius: 2em;
-      border: none;
-      margin: 0;
-      background-color: aliceblue;
-      padding: 20px 20px 0 20px;
-      box-shadow: 0px 0px 25px 1px black;
-   }
-}
-
-@media (min-width: 901px) and (min-height: 851px) {
-   #projectList {
-      display: flex;
-      flex-direction: row;
-      gap: 40px;
-      width: fit-content;
-      height: 45dvh;
-      padding: 10px;
-   }
-
-   .projectCard {
-      display: flex;
-      flex-direction: column;
-      flex-wrap: wrap;
-      flex-shrink: 0;
-      font-size: x-small;
-      width: clamp(250px, 50dvw, 400px);
-      height: 100%;
-      border-radius: 2em;
-      border: none;
-      background-color: aliceblue;
-      padding: 20px 20px 0 20px;
-      box-shadow: 0px 0px 25px 1px black;
-   }
-} */
-
-/* #content {
-   display: flex;
-   flex-direction: column;
-   width: fit-content;
-   height: fit-content;
-   overflow-x: scroll;
-} */
-
 #projectFilters {
    display: flex;
    flex-direction: column;
    justify-self: center;
    align-items: center;
    gap: 20px;
-   margin-top: 150px;
-   margin-bottom: 20px;
+   margin: 150px auto 50px auto;
    border-radius: 2.5em;
    height: 15dvh;
 
@@ -145,11 +81,11 @@
 .projectList {
    display: flex;
    flex-direction: row;
-   max-width: 80dvw;
-   overflow-x: scroll;
+   justify-content: center;
+   max-width: 90dvw;
+   flex-wrap: wrap;
    height: fit-content;
-   margin-left: auto;
-   margin-right: auto;
+   margin: 10px auto 10% auto;
    gap: 3em;
 
    > .projectCard {
