+++ styles/Shared.module.css
@@ -14,14 +14,14 @@
 
    > .skillCard {
       display: flex;
-      width: 12%;
+      width: min(12%, 100px);
       height: auto;
       aspect-ratio: 1/1;
       align-items: center;
       justify-content: center;
       border-radius: 5rem;
       padding: 1%;
-         
+
       > img {
          max-width: 80%;
          max-height: 80%;
